+++ CHANGELOG.md
@@ -1,14 +1,38 @@
-## [0.9.2] - 2024-01-10
-### Added
+### [0.9.2] - 2024-01-10
+
+#### Added
+
 - New payment verification feature.
-### Fixed
+
+#### Fixed
+
 - Bug in transaction handling.
 
-## [0.9.2] - 2024-01-01
-### Added
+### [0.9.2] - 2024-01-01
+
+#### Added
+
 - Initial release of Paylink Payment SDK.
 
-## [0.9.3] - 2024-01-08
-### Added
+### [0.9.3] - 2024-01-08
+
+#### Added
+
 - Added callback function to payment function.
 
+### [0.9.4] - 2024-03-06
+
+#### Added
+
+- API Link & Payment Frame Url based on project environment.
+- Payment token for authorization
+
+#### Fixed
+
+- Using Transaction Number & Order Number properly.
+
+### [1.0.0] - 2024-04-21
+
+#### Upgrade
+
+- Reimplement the whole SDK